feat: add exception handlers

+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -1,13 +1,16 @@
 package ru.yandex.practicum.filmorate.model;
 
-import jakarta.validation.constraints.*;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.PastOrPresent;
+import jakarta.validation.constraints.Pattern;
 import lombok.Data;
 
 import java.time.LocalDate;
 
 @Data
 public class User {
-    private Integer id;
+    private Long id;
 
     @NotBlank
     @Email
